+++ Assets/CommonCoreGame/RPGGame/GameParams.cs
@@ -30,13 +30,8 @@ namespace CommonCore
         public static PlayerViewType DefaultPlayerView { get; private set; } = PlayerViewType.PreferFirst;
 
         public static bool UseFallDamage { get; private set; } = true;
-        [Obsolete]
-        public static bool UseRandomDamage { get; private set; } = true;
         public static bool UseFriendlyFire { get; private set; } = true; //note that this can be overridden by player weapons and actor attack components
         public static float DamageFlashThreshold { get; private set; } = 0.01f; //very small because health is reported as a fraction
-        public static bool UseLegacyReequipShieldRecharge { get; private set; } = true; //use (bugged) Citadel behaviour for handling shield recharge state after item equip
-
-        public static bool AutoQuestNotifications { get; private set; } = true;
 
         public static bool AllowItemDropOutsideWorldScene { get; private set; } = false;
         public static bool AutocreateInventoryModels { get; private set; } = true; //if set, autocreates models for ammo and money